+++ tests/config.ts
@@ -1,18 +1,10 @@
 import { RuleTester } from 'eslint';
-
-// Override the types for RuleTester to allow for the parser property
-interface RuleTesterConfig {
-  parser?: string;
-  parserOptions?: object;
-}
+import parser from '@typescript-eslint/parser';
 
 export const ruleTester = new RuleTester({
-  parser: require.resolve('@typescript-eslint/parser'),
-  parserOptions: {
+  languageOptions: {
+    parser,
     ecmaVersion: 2020,
-    sourceType: 'module',
-    ecmaFeatures: {
-      jsx: true
-    }
+    sourceType: 'module'
   }
-} as any); 
+});